test: for every browser copyToClipboard
