+++ package.json
@@ -40,10 +40,10 @@
   "devDependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.18.6",
     "@babel/preset-typescript": "^7.18.6",
-    "@playwright/experimental-ct-react": "^1.23.3",
-    "@playwright/test": "^1.23.3",
-    "@types/react": "^18.0.15",
-    "eslint": "^8.19.0",
+    "@playwright/experimental-ct-react": "^1.24.2",
+    "@playwright/test": "^1.24.2",
+    "@types/react": "^18.0.16",
+    "eslint": "^8.21.0",
     "eslint-config-cheminfo-typescript": "^11.0.1",
     "prettier": "^2.7.1",
     "react": "^18.2.0",
