+++ src/copyToClipboard.ts
